+++ src/main/java/com/team9/anicare/schedule/controller/ScheduleController.java
@@ -1,7 +1,7 @@
 package com.team9.anicare.schedule.controller;
 
 
-import com.team9.anicare.common.Result;
+import com.team9.anicare.common.response.Result;
 import com.team9.anicare.schedule.dto.ScheduleDTO;
 import com.team9.anicare.schedule.service.ScheduleService;
 import org.springframework.beans.factory.annotation.Autowired;
