+++ src/main/java/com/team9/anicare/schedule/service/ScheduleService.java
@@ -1,8 +1,7 @@
 package com.team9.anicare.schedule.service;
 
-import com.team9.anicare.common.Result;
-import com.team9.anicare.common.ResultCode;
-import com.team9.anicare.pet.model.Pet;
+import com.team9.anicare.common.response.Result;
+import com.team9.anicare.common.exception.ResultCode;
 import com.team9.anicare.pet.repository.PetRepository;
 import com.team9.anicare.schedule.dto.ScheduleDTO;
 import com.team9.anicare.schedule.model.Schedule;
